+++ src/game/WorldHandlers/MovementHandler.cpp
@@ -288,7 +288,7 @@ void WorldSession::HandleMoveWorldportAckOpcode()
             }
 
             // mount capability changed
-            if (entry->ID != aura->GetModifier()->m_amount)
+            if (int32(entry->ID) != aura->GetModifier()->m_amount)
             {
                 if (MountCapabilityEntry const* oldEntry = sMountCapabilityStore.LookupEntry(aura->GetModifier()->m_amount))
                 {
