+++ src/game/WorldHandlers/SpellAuras.cpp
@@ -2941,7 +2941,7 @@ void Aura::PeriodicTick()
                 return;
             }
 
-            if (m_modifier.m_miscvalue < 0 || m_modifier.m_miscvalue >= MAX_POWERS)
+            if (m_modifier.m_miscvalue < 0 || m_modifier.m_miscvalue >= int32(MAX_POWERS))
             {
                 return;
             }
@@ -3083,7 +3083,7 @@ void Aura::PeriodicTick()
             DETAIL_FILTER_LOG(LOG_FILTER_PERIODIC_AFFECTS, "PeriodicTick: %s energize %s for %u dmg inflicted by %u",
                               GetCasterGuid().GetString().c_str(), target->GetGuidStr().c_str(), pdamage, GetId());
 
-            if (m_modifier.m_miscvalue < 0 || m_modifier.m_miscvalue >= MAX_POWERS)
+            if (m_modifier.m_miscvalue < 0 || m_modifier.m_miscvalue >= int32(MAX_POWERS))
             {
                 break;
             }
